fantasy-basketball/09-final-implementation/03-update-locked-points/requirements.txt
pandas 
requests 
openpyxl 
azure.identity 